Fix infeas cut correction, improve dwelltime cns handling, add ocp plots, mc (#14)
- Update infeasibility cut correction using gradient correction (sigma
correction deprecated)
- Enhancing OCP descriptor with flag for dwelltime constraints -- these
constraints should be considered only in the corresponding MIP and **not
when solving the NLP relaxation or the NLP with fixed integer
variables**.
- Using gurobi barrier method for MIQP root node relaxations for OCP
(`stcs` and `unstable_ocp`) - for `stcs` the barrier method takes 2
seconds, the dual simplex (default) takes >25 sec.
- Fix dwell time constraints in `stcs` and `unstable_ocp, now the
implementation coincides with the one in pycombina.
- Fix bug in `pycombina` when setting min up time 
- Fix bug when using the solver concatenation `cia+s-b-miqp`
- Add a script to create plots for OCPs (`stcs` and `unstable_ocp`)
- Add `stcs-simplified` which reduces the `stcs` OCP dimension by
considering a uniform grid with sampling time equal to 1 hour
- Some minor polishing